Split project loading code into new `project::load` module
